docs/ja_JP/SubmittingPatches: Add recommendation for mailing list references
Reflect the changes made in commit 9547c706d279 ("SubmittingPatches:
add recommendation for mailing list references").

Note 1: The URL of https://lkml.kernel.org/ still works and it is
kept as in the commit mentioned above.  Will be updated later.

Note 2: Omit a redundant modifier phrase of URL in the second
paragraph.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>