docs/conf.py: Work around imgmath regression of Sphinx 5.3.0
A regression related to the imgmath math-renderer extension was
reported [1], where math image to be embedded in html pages are
not properly located.  Although it has been fixed in Sphinx 6.0.1,
several dostros have picked 5.3.0 as their Sphinx package versions,
including Debian bookworm and Fedora 38 among others.

It can be worked around by setting "imgmath_embed = True" as
mentioned in the discussion in [1].

Set the variable conditionally for Sphinx 5.3.x.

Link: [1] https://github.com/sphinx-doc/sphinx/issues/10944
Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Cc: Mauro Carvalho Chehab <mchehab@kernel.org>