docs: Detect variable fonts and suggest removing them
Fedora and openSUSE has started deploying variable-font format
"Noto CJK" fonts [1, 2].
Unfortunately, XeTeX/XeLaTeX doesn't understand those fonts for
historical reasons and build of translations.pdf ends up in error
if such fonts are found on the build host.  To make matters worse,
Fedora enlists them in the dependencies of langpacks-zh_CN, -zh_TW,
-ko, and -ja packages.  This means those who have interest in CJK
(Chinese, Japanese, and Korean) pages in translations.pdf are more
likely to be hit by the error.

To help developers work around the issue, add a script for checking
existence of variable fonts and emitting suggestions.  Invoke it in
the error path of "make pdfdocs" so that it is activated only when
a PDF build actually fails.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
Reported-by: Jonathan Corbet <corbet@lwn.net>
Link: https://lore.kernel.org/r/8734tqsrt7.fsf@meer.lwn.net/
Reported-by: "Иван Иванович" <relect@bk.ru>
Link: https://lore.kernel.org/linux-doc/1708585803.600323099@f111.i.mail.ru/
---
This is not in line with Jon's suggestion of tweaking sphinx-pre-install
and/or kerneldoc-preamble.sty [1].

[1]: https://lore.kernel.org/r/87o7ccpcob.fsf@meer.lwn.net/

It would be ideal if kerneldoc-preamble.sty can be taught about the
way to detect the exsistence of variable fonts and to skip CJK contents.
However, it seems that it would add overly complex code for choosing
fonts by file names with meticulous options to select CJK variant (SC,
TC, KR, or JP) to use.  Furthermore, names of font files can be different
among distro packages and those manually installed from the Noto CJK font
site.

As for sphinx-pre-install, even if variable "Noto CJK" fonts are found,
affected PDF is translations.pdf only (at the moment).  So emitting warning
every time would be noisy.

Therefore, as a minimal approach, to help develpers find a way forward,
giving suggestions after the fact sounds like a reasonable approach to me,
at least for the time being.

Jon, you said in [1]:

>> What does
>>
>>     fc-list | grep NotoSansCJK-VF.ttc
>>
>> say?
>
> No output at all, even though I had google-noto-sans-cjk-vf-fonts
> installed.

So I'm wondering if scripts/check-variable-font.sh in this patch emits
helpful suggestions.  Can you give it a try?

  - Install google-noto-sans-cjk-vf-fonts
  - Install google-noto-serif-cjk-vf-fonts
  - What does "sh scripts/check-variable-font.sh" say?
  - Run "make cleandocs; make SPHINXDIRS=translations pdfdocs".
    Do you see the suggestion after the build error?
  - Uninstall above 2 font packages.
  - What does "sh scripts/check-variable-font.sh" say now?
  - Run "make cleandocs; make SPHINXDIRS=translations pdfdocs".
    Does it complete successfully?

        Thanks, Akira