docs/ja_JP: Use "admonition" directive in stead of "note"
It would be a better approach to change the way "note" directive
is rendered in HTML or PDF output so that it looks natural in
ja_JP translation. However, current Sphinx does not provide a
means to change language conventions in the middle of the whole
documentation tree.

As a compromise, use the low level directive "admonition" in place
of the "note" directive and try to achieve the natual looks of those
notes in Japanese.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>
---
This might well be regarded as "over-engineering", but as we have
very few reST docs under ja_JP/, this can be in the range of
maintanable ...