docs: kerneldoc-preamble: Set \footnotesize for literal and code blocks
Combination of modern Sphinx and LaTeX allows long lines in literal
(verbatim) blocks to be automatically wrapped around in PDF output.
This is good for preventing them from pushed away fron right-side
margins, but results in ugly and hard-to-parse code-snippets,
ascii-diagrams, log messages, and so on.

There are adhoc tweaks using ".. raw:: latex" and specifying
\footnotesize LaTeX macro case-by-case basis under Documentation/.

Instead, add needed settings in our LaTeX preamble and make
\footnotesize as the default size for literal block and the like
everywhere.

For special cases where a different font size is preferred, use
\setlitealfontsize{} macro defined here, e.g.:

.. raw:: latex

    \begingroup
    \setliteralfontsize{\scriptsize}

    [literal blocks such as ".. include:: + :literal:",
    ".. literalinclude::", ".. code-block::", and so on.]

.. raw:: latex

    \endgroup

Grouping is recommended for the fontsize selection is local to the
literal contents at hand.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>