docs: kerneldoc-preamble: Set \footnotesize for literal and code blocks
Combination of modern Sphinx and LaTeX allows long lines in literal
(verbatim) blocks to be automatically wrapped around in PDF output.
This is good for preventing them from pushed beyond right-side margins,
but results in ugly and hard-to-parse code-snippets, ascii-diagrams,
log messages, and so on.

There are adhoc tweaks using ".. raw:: latex" and specifying
\footnotesize LaTeX macro case-by-case basis under Documentation/.

Instead, add needed settings in our LaTeX preamble and make
\footnotesize as the default size for literal block and the like
everywhere.

Signed-off-by: Akira Yokosawa <akiyks@gmail.com>