scripts: sphinx-build-wrapper: add a wrapper for sphinx-build
There are too much magic inside docs Makefile to properly run
sphinx-build. Create an ancillary script that contains all
kernel-related sphinx-build call logic currently at Makefile.

Such script is designed to work both as an standalone command
and as part of a Makefile. As such, it properly handles POSIX
jobserver used by GNU make.

It should be noticed that, when running the script alone,
it will only take care of sphinx-build and cleandocs target.
As such:

- it won't run "make rustdoc";
- no extra checks.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>