docs: kernel_include.py: Remove mostly unused Include class
While the original code came from the Sphinx Include class,
such class is monolithic: it has only one function that does
everything, and 3 variables that are used:

	- required_arguments
	- optional_arguments
	- option_spec

Those are the same vars that every other Sphinx directive
extension has to define. Also, kernel-include doesn't support
the current set of option_spec.

So, pick what it is supported, dopping the rest.

Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>