gfs2: Remove useless transaction in evict_linked_inode
In evict_linked_inode(), the truncate_inode_pages() calls are carried
out inside a transaction.  This code was added to what was then function
gfs2_delete_inode() in commit 16615be18cadf ("[GFS2] Clean up journaled
data writing").

It's unclear what that transaction was originally good for, but for
unlinked inodes, page cache truncation is meanwhile handled in the hole
punching code and no longer in evict_unlinked_inode().  For linked
inodes, there is no need for a transaction, so remove that code from
evict_linked_inode().

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>